+++ tests/test_checker/test_noqa.py
@@ -121,6 +121,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z423': 1,
         'Z424': 1,
         'Z425': 1,
+        'Z426': 1,
 
         'Z430': 1,
         'Z431': 2,